refactor: dataモデルを別ファイルに分割
