+++ app/schemas/api_response.py
@@ -1,11 +1,9 @@
-from pydantic import BaseModel, create_model, Field
+from pydantic import BaseModel
 
 from app.models.character_usage import CharacterUsages
-from app.models.settings import GuildSettings
-from app.models.subscription import Subscription
 from app.models.trusted_roles import TrustedRoles
+from app.schemas.api_data import ConnectionStateData, DictData, SubscriptionsData
 from app.schemas.connection_command import ConnectionCommand
-from app.schemas.connection_state import ConnectionState
 from app.schemas.message_link_expand_pref import MessageLinkExpansionPreference
 
 
@@ -23,19 +21,9 @@ class ConnectionCommandAPIResponse(BaseModel):
     data: ConnectionCommand
 
 
-# todo: connection_state(単数形)に修正
-ConnectionStateData = create_model("ConnectionStateDataModel", connection_states=(ConnectionState, ...))
-
-
 class ConnectionStateAPIResponse(BaseModel):
     message: str
     data: ConnectionStateData
-
-
-DictData = create_model(
-    "DictDataModel",
-    dict_=(dict, Field(..., alias="dict"))
-)
 
 
 class DictAPIResponse(BaseModel):
@@ -49,15 +37,9 @@ class MessageLinkExpandAPIResponse(BaseModel):
     data: MessageLinkExpansionPreference
 
 
-GuildSettingsData = create_model("GuildSettingsDataModel", settings=(GuildSettings, ...))
-
-
 class GuildSettingsAPIResponse(BaseModel):
     message: str
     data: dict
-
-
-SubscriptionsData = create_model("SubscriptionsDataModel", subscriptions=(list[Subscription], ...))
 
 
 class SubscriptionAPIResponse(BaseModel):
